create a simple executor type
in the rust async world, an executor is the middleware
between blocking and non-blocking code.
It is responsible for orchestrating tasks to run, as close to completion as possible.
It is also responsible for switching between tasks, managing poling
strategies, etc.
Because the executor needs to keep track of every spawned task it manages,
it's obvious some kind of data structure is needed.
For this implementation, I used the doubly ended queue from the `alloc` crate.
I used this particular data structure because it allows me
to push tasks on the end of the queue with spawn, while being able to take the next one to pole,
based on a fifo(first in, first out) principle
In the future, this queue might need to be changed due to syncronisation issues, but for now this should be good enough