begin a skech of the waker type
An async executor poles each future in its queue, untill all of them reached completion.
There are two ways of poling futures, the naive way and the efficient way
the naive implementation consists of
a simple loop inside the run method of the executor,
which poles all futures repeatedly, not adding them to the front
of the queue once they returned `Pole::Ready`
The efficient way involves making the executor
sleep for most of the time in that infinite loop,
only waking futures that
signal they need to be woken because they can make progress. This is accomplished through the waker type
For now, this waker does nothing, that's why it's constructed unsafely from a raw waker vtable.
It does nothing because we only need it for constructing the context argument that will be passed to
each future on the queue, where futures usually store their own wakers in other executor implementations.