add async method to spawn tasks on the executor
Before, separate tasks could only be spawned
on the executor from a non-async context,
for example the main function. However, async tasks also have to
be able to spawn their own concurrent units of work, so an async method is created to support this use case.
For now, this method is the same as the normal spawn,
since it should return quick enough for the executor to not be blocked for more than afew microseconds.